refactor: consolidate AI endpoints
