+++ docs/api-overview.md
@@ -5,10 +5,7 @@ All API routes are under `api/` and are deployed as serverless functions.
 | Route | Description |
 |-------|-------------|
 | `api/access-key` | Validates invitation keys and marks them as used. |
-| `api/estimate-cost` | Uses OpenAI to estimate the total cost of a recipe. Requires authentication. |
-| `api/generate-description` | Generates a short recipe description with OpenAI. |
-| `api/format-instructions` | Formats raw cooking instructions into clear steps using OpenAI. Requires authentication. |
-| `api/generate-image` | Generates a DALL·E image, uploads it to Supabase and returns the file path. |
+| `api/ai` | Handles AI tasks like description generation, image creation, cost estimation and instruction formatting based on the `action` query parameter (`description`, `image`, `cost`, `format`). Requires authentication. |
 | `api/update-profile` | Updates fields in `public_user_view` for the current user. |
 | `api/purchase-credits` | Starts a Stripe Checkout session to buy extra AI credits. |
 | `api/buy-ia-credits` | Alias for `api/purchase-credits`. |
